Got the project going with some styling
Changed some colors and font-family. I also changed some values for the hr tag, h1, h2, a, a:hover, a:active, and a:focus. I made some classes, most importantly the rounded-container class! It allows the body to be in it's own rounded cornered box.   For more of the changes look at the changes in the code.